+++ example/client/client1/components/app/index.js
@@ -1,22 +1,21 @@
 import React, { Component } from 'react'
 import CatalogComponent, { withCatalog } from 'react-component-catalog'
 
-let Button
-
 class App extends Component {
   constructor(props) {
     super(props)
 
     // you can import and use registered components with catalog.getComponent
     const { catalog } = props
-    Button = catalog.getComponent('Button')
+    this.Button = catalog.getComponent('Button')
   }
 
   render() {
+    const Button = this.Button
     // or you use them with the <CatalogComponent /> component
     return (
       <div>
-        <CatalogComponent name="Title">Hello Client1</CatalogComponent>
+        <CatalogComponent component="Title">Hello Client1</CatalogComponent>
         <Button />
       </div>
     )