+++ example/client/client1/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 
 import { CatalogProvider } from 'react-component-catalog'
-import catalog, { App } from './components'
+import catalog, { App } from './catalog'
 
 ReactDOM.render(
   <CatalogProvider catalog={catalog}>