+++ jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  collectCoverageFrom: ['src/**/*.js'],
+  setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
+  testPathIgnorePatterns: ['<rootDir>/(dist|node_modules)/'],
+  transform: {
+    '^.+\\.js$': 'babel-jest',
+  },
+}