+++ src/components/catalog-component.js
@@ -1,10 +1,24 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 import withCatalog from './with-catalog'
 
 class CatalogComponent extends React.Component {
+  static propTypes = {
+    catalog: PropTypes.object.isRequired,
+    component: PropTypes.string.isRequired,
+  }
+
   render() {
-    const { catalog, name, ...others } = this.props
-    const Component = catalog.getComponent(name)
+    const { catalog, component, ...others } = this.props
+
+    if (!catalog) {
+      return null
+    }
+    const Component = catalog.getComponent(component)
+
+    if (!Component) {
+      return null
+    }
     return <Component {...others} />
   }
 }