+++ src/components/with-catalog.test.js
@@ -1,0 +1,67 @@
+import React, { Component } from 'react'
+import { mount } from 'enzyme'
+
+import Catalog from '../lib/catalog'
+import withCatalog from './with-catalog'
+
+const TestComponent = withCatalog(() => <div>Hello World</div>)
+
+const TestClassComponent = withCatalog(
+  class TestClassComponent extends Component {
+    render() {
+      return <div>Hello Class</div>
+    }
+  },
+)
+
+describe('withCatalog', () => {
+  let testCatalog
+  let testContext
+
+  beforeEach(() => {
+    testCatalog = new Catalog({
+      components: {
+        TestClassComponent,
+        TestComponent,
+      },
+    })
+
+    testContext = {
+      context: { catalog: testCatalog },
+    }
+  })
+
+  it('renders a wrapped functional component properly', () => {
+    const wrapper = mount(<TestComponent hello="world" />, testContext)
+
+    // with proper displayName
+    expect(wrapper.name()).toEqual('withCatalog(withCatalogChild)')
+
+    // passed props and the context as a new prop
+    expect(wrapper.childAt(0).props()).toEqual({
+      catalog: testCatalog,
+      hello: 'world',
+    })
+
+    // and the component itself
+    expect(wrapper.find(TestComponent).prop('hello')).toEqual('world')
+    expect(wrapper.find(TestComponent).text()).toEqual('Hello World')
+  })
+
+  it('renders a wrapped react class component properly', () => {
+    const wrapper = mount(<TestClassComponent hello="world" />, testContext)
+
+    // with proper displayName
+    expect(wrapper.name()).toEqual('withCatalog(TestClassComponent)')
+
+    // passed props and the context as a new prop
+    expect(wrapper.childAt(0).props()).toEqual({
+      catalog: testCatalog,
+      hello: 'world',
+    })
+
+    // and the component itself
+    expect(wrapper.find(TestClassComponent).prop('hello')).toEqual('world')
+    expect(wrapper.find(TestClassComponent).text()).toEqual('Hello Class')
+  })
+})